Extract the error module of LibOS as a separate crate
