+++ src/libos/Cargo.toml
@@ -12,6 +12,7 @@ async-rt  = { path = "crates/async-rt", features = ["sgx"] }
 atomic = "0.5"
 bitflags = "1.0"
 bitvec = { version = "0.17", default-features = false, features = ["alloc"]  }
+errno = { path = "crates/errno", features = ["occlum"] }
 log = "0.4"
 aligned = "0.3.4"
 lazy_static = { version = "1.1.0", features = ["spin_no_std"] } # Implies nightly
