+++ src/libos/src/lib.rs
@@ -67,8 +67,6 @@ use crate::process::pid_t;
 
 #[macro_use]
 mod prelude;
-#[macro_use]
-mod error;
 
 mod config;
 mod entry;
