+++ src/libos/src/prelude.rs
@@ -13,12 +13,13 @@ pub use std::sync::{
     Arc, SgxMutex, SgxMutexGuard, SgxRwLock, SgxRwLockReadGuard, SgxRwLockWriteGuard,
 };
 
-// Override prelude::Result with error::Result
-pub use crate::error::Result;
-pub use crate::error::*;
 pub use crate::fs::{File, FileDesc, FileRef};
 pub use crate::process::{pid_t, uid_t};
 pub use crate::util::sync::RwLock;
+pub use errno::prelude::*;
+
+// To override the default Result type
+pub use errno::prelude::Result;
 
 macro_rules! debug_trace {
     () => {
@@ -44,4 +45,32 @@ pub fn align_down(addr: usize, align: usize) -> usize {
 
 pub fn unbox<T>(value: Box<T>) -> T {
     *value
+}
+
+// return Err(errno) if libc return -1
+macro_rules! try_libc {
+    ($ret: expr) => {{
+        let ret = unsafe { $ret };
+        if ret < 0 {
+            let errno = unsafe { libc::errno() };
+            return_errno!(Errno::from(errno as u32), "libc error");
+        }
+        ret
+    }};
+}
+
+// return Err(errno) if libc return -1
+// raise SIGPIPE if errno == EPIPE
+macro_rules! try_libc_may_epipe {
+    ($ret: expr) => {{
+        let ret = unsafe { $ret };
+        if ret < 0 {
+            let errno = unsafe { libc::errno() };
+            if errno == Errno::EPIPE as i32 {
+                crate::signal::do_tkill(current!().tid(), crate::signal::SIGPIPE.as_u8() as i32);
+            }
+            return_errno!(Errno::from(errno as u32), "libc error");
+        }
+        ret
+    }};
 }
